some apis are secured

+++ src/main/java/com/team/updevic001/services/impl/CommentServiceImpl.java
@@ -8,15 +8,19 @@ import com.team.updevic001.dao.entities.User;
 import com.team.updevic001.dao.repositories.CommentRepository;
 import com.team.updevic001.dao.repositories.CourseRepository;
 import com.team.updevic001.dao.repositories.LessonRepository;
+import com.team.updevic001.exceptions.ForbiddenException;
 import com.team.updevic001.exceptions.ResourceNotFoundException;
 import com.team.updevic001.model.dtos.request.CommentDto;
 import com.team.updevic001.model.dtos.response.comment.ResponseCommentDto;
 import com.team.updevic001.services.CommentService;
+import com.team.updevic001.utility.AuthHelper;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 
+@Slf4j
 @Service
 @RequiredArgsConstructor
 public class CommentServiceImpl implements CommentService {
@@ -24,30 +28,30 @@ public class CommentServiceImpl implements CommentService {
     private final CourseRepository courseRepository;
     private final CommentMapper commentMapper;
     private final LessonRepository lessonRepository;
-    private final UserServiceImpl userServiceImpl;
     private final CommentRepository commentRepository;
+    private final AuthHelper authHelper;
 
     @Override
-    public void addCommentToCourse(String userId, String courseId, CommentDto commentDto) {
-        User user = userServiceImpl.findUserById(userId);
+    public void addCommentToCourse(String courseId, CommentDto commentDto) {
+        User authenticatedUser = authHelper.getAuthenticatedUser();
         Course course = courseRepository.findById(courseId)
                 .orElseThrow(() -> new ResourceNotFoundException("There is no such course."));
         Comment comment = Comment.builder()
                 .content(commentDto.getContent())
-                .user(user)
+                .user(authenticatedUser)
                 .course(course)
                 .build();
         commentRepository.save(comment);
     }
 
     @Override
-    public void addCommentToLesson(String userId, String lessonId, CommentDto commentDto) {
-        User user = userServiceImpl.findUserById(userId);
+    public void addCommentToLesson(String lessonId, CommentDto commentDto) {
+        User authenticatedUser = authHelper.getAuthenticatedUser();
         Lesson lesson = lessonRepository.findById(lessonId)
                 .orElseThrow(() -> new ResourceNotFoundException("There is no such lesson."));
         Comment comment = Comment.builder()
                 .content(commentDto.getContent())
-                .user(user)
+                .user(authenticatedUser)
                 .lesson(lesson)
                 .build();
         commentRepository.save(comment);
@@ -55,8 +59,13 @@ public class CommentServiceImpl implements CommentService {
 
     @Override
     public void updateComment(String commentId, CommentDto commentDto) {
+        User authenticatedUser = authHelper.getAuthenticatedUser();
         Comment comment = commentRepository.findById(commentId)
-                .orElseThrow(() -> new ResourceNotFoundException("This comment has been deleted."));
+                .orElseThrow(() -> new ResourceNotFoundException("COMMENT_NOT_FOUND"));
+        if (!comment.getUser().getUuid().equals(authenticatedUser.getUuid())) {
+            log.error("User wit ID {} not allowed to delete comment with ID {}: User is not author of comment", authenticatedUser.getUuid(), commentId);
+            throw new ForbiddenException("NOT_ALLOWED_UPDATE_COMMENT");
+        }
         comment.setContent(commentDto.getContent());
         commentRepository.save(comment);
     }
@@ -79,6 +88,16 @@ public class CommentServiceImpl implements CommentService {
 
     @Override
     public void deleteComment(String commentId) {
+        User authenticatedUser = authHelper.getAuthenticatedUser();
+        Comment comment = commentRepository.findById(commentId)
+                .orElseThrow(() -> new ResourceNotFoundException("COMMENT_NOT_FOUND"));
+        if (!comment.getUser().getUuid().equals(authenticatedUser.getUuid())
+                || comment.getCourse().getTeacherCourses().stream()
+                .map(tc -> tc.getTeacher().getUser().getUuid())
+                .noneMatch(authenticatedUser.getUuid()::equals)) {
+            log.error("User with ID {} not allowed to delete comment: User must be either admin of the course or author of course", authenticatedUser.getUuid());
+            throw new ForbiddenException("NOT_ALLOWED_DELETE_COMMENT");
+        }
         commentRepository.deleteById(commentId);
     }
 }
